requirements.txt
fastapi>=0.110.1
motor>=3.3.2
uvicorn>=0.25.0
supabase>=2.4.5
redis>=5.0.4
boto3>=1.34.129
requests-oauthlib>=2.0.0
cryptography>=42.0.8
python-dotenv>=1.0.1
litellm>=1.52.3
langsmith>=0.1.142
kubernetes==29.0.0
paramiko==3.4.0
tenacity==8.2.3
python-json-logger==2.0.7
pytest==8.0.0
pytest-cov==4.1.0
black==24.1.1
flake8==7.0.0
mypy==1.8.0
pyyaml>=6.0.2
prometheus-client==0.19.0
structlog==24.1.0
typing-extensions>=4.12.2
google-cloud-pubsub>=2.26.1
ghapi>=1.0.6
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.10
pydantic>=2.9.2
pytest-mock>=3.14.0
typer>=0.14.0
requests>=2.31.0
gitpython>=3.1.44
setuptools>=45
wheel
